+++ src/top/Test1_calculate.java
@@ -24,9 +24,9 @@ public class Test1_calculate {
         for (int i = 1; i <= 1281; i++) {
             x0.add((double) i);
         }
-        List<Integer> y0 = new ArrayList<Integer>();
+        List<Double> y0 = new ArrayList<Double>();
         for (int i = 1; i <= 1280; i++) {
-            y0.add(1);
+            y0.add(1.0);
         }
         Column_ian ck0 = new Column_ian(step0, x0, y0);
         CKdist.add(ck0);
@@ -36,9 +36,9 @@ public class Test1_calculate {
         for (int i = 1; i <= 6; i++) {
             x1.add((double) i);
         }
-        List<Integer> y1 = new ArrayList<Integer>();
+        List<Double> y1 = new ArrayList<Double>();
         for (int i = 1; i <= 5; i++) {
-            y1.add(1);
+            y1.add(1.0);
         }
         for (int i = 1; i < 8; i++) {
             Column_ian ck1 = new Column_ian(step1, x1, y1);
@@ -50,7 +50,7 @@ public class Test1_calculate {
 //        for (int i = 1; i <= 11; i++) {
 //            x1.add((double) i);
 //        }
-//        List<Integer> y1 = new ArrayList<Integer>();
+//        List<Double> y1 = new ArrayList<Double>();
 //        for (int i = 1; i <= 10; i++) {
 //            y1.add(1);
 //        }
